requirements.txt
torch==2.4.0
numpy==2.0.1
pillow==10.4.0
av==12.3.0
streamlit==1.37.1
matplotlib==3.9.1.post1
plotly==5.23.0
tqdm==4.66.5
pyyaml==6.0.2